backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0
aiofiles==23.2.1
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
openai==1.3.0
